Bug fix in precalc - initial excitation generators defined as REAL*8 rather than integer.
More work done on GraphMorph - nearly done.


git-svn-id: https://wwmm.ch.cam.ac.uk/svn2/groups/alavi/NECI/trunk@8805 3c9ecaa4-0e0d-0410-a8e4-acb18276dcea

+++ precalc.F90
@@ -7,7 +7,7 @@ module PreCalc
     
     Use Integrals, only : ChemPot
     REAL*8, POINTER, DIMENSION(:) :: PGENLIST
-    REAL*8, POINTER, DIMENSION(:) :: NMEM
+    INTEGER, POINTER, DIMENSION(:) :: NMEM
     REAL*8, POINTER, DIMENSION(:,:) :: GRAPHPARAMS
     INTEGER, POINTER, DIMENSION(:,:,:) :: GRAPHS
 #if defined(POINTER8)
